Lay groundwork for supporting flags + don't clear db
Functionality is beginning to be refactored into separate methods and
files here. In particular, I'm close to adding CLI flag
configurability (that is, the user will for now be able to select bus
lines they'd like to add to, or remove from, the database.)

The one "breaking change" here is that the database is no longer
cleared on each call. This is simply because we want our added bus
stops to persist across calls now.