Add all routes provided to -a/--add flag
There is a problem: currently, the id is a primary key, and so has a
uniqueness constraint. However, I need the id-routeID pair to have the
uniqueness constraint (various routes can share the same bus stop!)

And so, I need to think of another primary key, perhaps simply an
integer increment.