goose: revise schema to specify uniqueness by route+stop
In other words, we were mistakenly using the stop id (which has been
renamed appropriately from the less specific "id") as the primary key;
however, multiple buses can use the same stop. For example, the M11
and the M12 share a few stops in common; and so, adding the M12 stops
after having added the M11 stops would result in a constraint
error. This has now been fixed.