refactor: make duplicated CLI arg detection more robust
The flags.go helper logic checks for this as soon as it happens, as
opposed to waiting until actual adds and removes are attempted. IOW,
what happened previously was that, even though an argument might've
been duplicated (e.g. -a M11 -r M11), since adding occurs before
removing in the code itself, the attempt to add the M11 route would've
still succeeded, which is undesirable.